sidebar optimisation, cleared render error, updated modal bg color
