+++ src/components/Sidebar/Sidebar.jsx
@@ -15,43 +15,24 @@ import { BONDS } from "../../constants";
 import { useMediaQuery } from "@material-ui/core";
 
 
-function Sidebar({ isExpanded, setRoute, address, provider, blockExplorer, theme }) {
+function Sidebar({ isExpanded, theme, ohmDaiBondDiscount, ohmFraxLpBondDiscount, daiBondDiscount, currentIndex }) {
   const dispatch = useDispatch();
-  const [isActive, setIsActive] = useState("stake");
-  const currentBlock  = useSelector((state ) => { return state.app.currentBlock });
-  const currentIndex = useSelector((state) => { return state.app.currentIndex });
-
-
-  const daiBondDiscount = useSelector(state => {
-    return state.bonding['dai'] && state.bonding['dai'].bondDiscount;
-  });
-
-  const ohmDaiBondDiscount = useSelector(state => {
-    return state.bonding['ohm_dai_lp'] && state.bonding['ohm_dai_lp'].bondDiscount;
-  });
-
-  const ohmFraxLpBondDiscount = useSelector(state => {
-    return state.bonding['ohm_frax_lp'] && state.bonding['ohm_frax_lp'].bondDiscount;
-  })
+  const [isActive, setIsActive] = useState();
 
   const checkPage = useCallback((match, location, page) => {
     console.log(match);
     const currentPath = location.pathname.replace("/", "");
-
     const currentURL = window.location.href;
 
     if (currentPath.indexOf("dashboard") >= 0 && page === "dashboard") {
-      setIsActive("dashboard");
       return true;
     }
 
     if (currentPath.indexOf("stake") >= 0 && page === "stake") {
-      setIsActive("stake");
       return true;
     }
 
     if ((currentPath.indexOf("bonds") >= 0 || currentPath.indexOf("choose_bond") >= 0) && page === "bonds") {
-      setIsActive("bonds");
       return true;
     }
 
@@ -76,17 +57,17 @@ function Sidebar({ isExpanded, setRoute, address, provider, blockExplorer, theme
 
         <div className="dapp-menu-links">
           <div className="dapp-nav" id="navbarNav">
-            <NavLink onClick={() => { setRoute("/dashboard" ) }} to="/dashboard" className={`button button-dapp-menu ${isActive == "dashboard" && "active"}`} isActive={(match, location) => { return checkPage(match, location, "dashboard") }}>
+            <NavLink to="/dashboard" isActive={(match, location) => { return checkPage(match, location, "dashboard") }} className={`button button-dapp-menu ${isActive && "active"}`}>
               <DashboardIcon className="me-3" />
               <span>Dashboard</span>
             </NavLink>
 
-            <NavLink onClick={() => { setRoute("/" ) }} to="/" className={`button button-dapp-menu ${isActive == "stake" && "active"}`} isActive={(match, location) => { return checkPage(match, location, "stake") }}>
+            <NavLink to="/" isActive={(match, location) => { return checkPage(match, location, "stake") }}  className={`button button-dapp-menu ${isActive && "active"}`} >
               <StakeIcon className="me-3" />
               <span>Stake</span>
             </NavLink>
 
-            <NavLink onClick={() => { setRoute("/bonds" ) }} to="/bonds" className={`button button-dapp-menu ${isActive == "bond" && "active"}`} isActive={(match, location) => { return checkPage(match, location, "bonds") }}>
+            <NavLink to="/bonds" isActive={(match, location) => { return checkPage(match, location, "bonds") }} className={`button button-dapp-menu ${isActive && "active"}`}>
               <BondIcon className="me-3" />
               <span>Bond</span>
             </NavLink>
@@ -98,7 +79,7 @@ function Sidebar({ isExpanded, setRoute, address, provider, blockExplorer, theme
                 <p>OHM-DAI LP<span>{trim(ohmDaiBondDiscount * 100, 2)}%</span></p>
                 <p>OHM-FRAX LP<span>{trim(ohmFraxLpBondDiscount * 100, 2)}%</span></p>
                 <p>DAI<span>{trim(daiBondDiscount * 100, 2)}%</span></p>
-                {/* <p>FRAX<span></span></p> */}
+                <p>FRAX<span></span></p>
               </div>
             </div>
           </div>
@@ -108,7 +89,7 @@ function Sidebar({ isExpanded, setRoute, address, provider, blockExplorer, theme
 
         <div className="dapp-menu-external-links">
           { Object.keys(externalUrls).map((link, i) => {
-            return <a key={i} href={externalUrls[link].url} target="_blank" className="button button-dapp-menu">
+            return <a key={i} href={`${externalUrls[link].url}`} target="_blank" className="button button-dapp-menu">
               {externalUrls[link].icon}
               <span>{externalUrls[link].title}</span>
             </a>
