+++ src/theme.js
@@ -2,7 +2,7 @@ export const darkTheme = {
     color: "#ffffff",
     backgroundColor: "#3A4050",
     background: "radial-gradient(circle at 25% 0%, rgba(227,255,240,.5), rgba(227,255,240,0) 50%), radial-gradient(circle at 80% 80%, rgba(131,165,203,.5), rgba(131,165,203,0) 50%)",
-    cardBg: "rgba(27, 29, 34, 0.2)",
+    cardBg: "rgba(27, 29, 34, 0.4)",
     modalBg: "rgba(27, 29, 34, 0.8)",
     cardBorder: "none",
     logoColor: "invert(1)",
