+++ src/views/Bond/bond.scss
@@ -8,7 +8,7 @@
         .ohm-card {
             // background-color: #1B1D22FF !important;
             opacity: 1 !important;  
-            background-color: rgba(27,29,34,0.9) !important;
+            // background-color: rgba(27,29,34,0.9) !important;
             height: 55vh;
         }
     }
